planner class refactor
