+++ docs/temp_data_gen.md
@@ -50,7 +50,7 @@ To create a quadruple of python code, code description, PDDL, and Plan, there sh
 import unittest
 from nl2flow.compile.flow import Flow
 from nl2flow.compile.operators import ClassicalOperator as Operator
-from nl2flow.plan.planners import Michael, Christian
+from nl2flow.plan.planner import Michael, Christian
 from nl2flow.plan.options import DEFAULT_PLANNER_URL
 from nl2flow.compile.schemas import SignatureItem, GoalItem, GoalItems
 from profiler.test_helpers.profiler_test_helper import write_pddl_plan
