+++ tests/debugger/test_debugger_from_flow.py
@@ -4,7 +4,7 @@ from nl2flow.compile.flow import Flow
 from nl2flow.compile.operators import ClassicalOperator as Operator
 from nl2flow.compile.schemas import SignatureItem, Parameter, Constraint, GoalItems, GoalItem, Step
 from nl2flow.compile.options import LifeCycleOptions, BasicOperations
-from nl2flow.plan.planners import Kstar
+from nl2flow.plan.planners.kstar import Kstar
 from copy import deepcopy
 
 PLANNER = Kstar()
