+++ tests/planner/formatting/test_prettifier.py
@@ -1,4 +1,4 @@
-from nl2flow.plan.planners import Kstar
+from nl2flow.plan.planners.kstar import Kstar
 from nl2flow.compile.flow import Flow
 from nl2flow.compile.operators import ClassicalOperator as Operator
 from nl2flow.compile.schemas import SignatureItem, Parameter, Constraint, GoalItems, GoalItem
