+++ tests/profiler/generators/dataset_generator/test_dataset_generator.py
@@ -1,4 +1,4 @@
-from nl2flow.plan.planners import Kstar
+from nl2flow.plan.planners.kstar import Kstar
 from profiler.data_types.generator_data_type import AgentInfoGeneratorInput
 from profiler.generators.dataset_generator.dataset_generator import (
     generate_dataset_with_info_generator,
