+++ tests/profiler/generators/quadruple_generator/test_quadruple_generator.py
@@ -1,6 +1,6 @@
 from nl2flow.compile.flow import Flow
 from nl2flow.compile.operators import ClassicalOperator as Operator
-from nl2flow.plan.planners import Kstar
+from nl2flow.plan.planners.kstar import Kstar
 from nl2flow.compile.schemas import SignatureItem, GoalItem, GoalItems
 from profiler.test_helpers.profiler_test_helper import write_pddl_plan
